requirements.txt
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
Click==7.0
cssmin==0.2.0
decorator==4.4.2
dominate==2.4.0
Flask==1.1.1
Flask-Assets==2.0
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-nav==0.6
Flask-Principal==0.4.0
Flask-Security-Too==3.3.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
jsmin==2.2.2
-e git+https://github.com/louking/loutilities.git@68cabc4e2136e2805f8166ad0d65d610b1e51c79#egg=loutilities
MarkupSafe==1.1.1
mysqlclient==1.4.6
passlib==1.7.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.13
sqlalchemy-datatables==2.0.1
validators==0.14.2
visitor==0.1.3
webassets==2.0
Werkzeug==1.0.0
WTForms==2.2.1
